Fix FPS dependant scroll in JumpyJet sample (and some typos)

+++ samples/Games/JumpyJet/JumpyJet.Game/BackgroundSection.cs
@@ -31,7 +31,7 @@ namespace JumpyJet
         public bool IsUpdating { get; set; }
         public bool IsRunning { get; protected set; }
         public bool IsVisible { get; protected set; }
-        public int ScrollPos { get; protected set; }
+        public float ScrollPos { get; protected set; }
         public float ScrollWidth { get; protected set; }
         public float ScrollSpeed { get; protected set; }
 
@@ -60,7 +60,7 @@ namespace JumpyJet
                 if (ScrollPos > textureRegion.Width)
                     ScrollPos = 0;
 
-                ScrollPos += (int)(elapsedTime * ScrollSpeed);
+                ScrollPos += elapsedTime * ScrollSpeed;
 
                 UpdateSpriteQuads();
             }
